requirements.txt
black==23.3.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
clickhouse-connect==0.5.25
DateTime==5.1
Deprecated==1.2.14
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Limiter==3.3.1
Flask-Session==0.5.0
gunicorn==20.1.0
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.24.3
ordered-set==4.1.0
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
platformdirs==3.5.3
pluggy==1.0.0
Pygments==2.15.1
pytest==7.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rich==13.3.5
six==1.16.0
structlog==23.1.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
wrapt==1.15.0
zope.interface==6.0
zstandard==0.21.0
